+++ src/components/pages/ProblemDetailPage/ProblemDetailPage.style.tsx
@@ -217,33 +217,3 @@ export const ModalInnerText = styled.p<{ color?: string }>`
   font-weight: bold;
   color: ${({ color }) => color || BLACK};
 `;
-
-export const BookMarkIcon = ({
-  isFavorite,
-  onClick,
-}: {
-  isFavorite: boolean;
-  onClick: () => void;
-}) => (
-  <button
-    type="button"
-    onClick={onClick}
-    aria-label={isFavorite ? '즐겨찾기 제거' : '즐겨찾기 추가'}
-  >
-    <svg
-      width="20"
-      height="26"
-      viewBox="0 0 20 26"
-      fill={isFavorite ? PRIMARY_PURPLE : 'none'}
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M1.25 7.75C1.25 5.6498 1.25 4.5997 1.65873 3.79754C2.01825 3.09193 2.59193 2.51825 3.29754 2.15873C4.0997 1.75 5.1498 1.75 7.25 1.75H12.75C14.8502 1.75 15.9003 1.75 16.7025 2.15873C17.4081 2.51825 17.9817 3.09193 18.3413 3.79754C18.75 4.5997 18.75 5.6498 18.75 7.75V24.25L10 19.25L1.25 24.25V7.75Z"
-        stroke={isFavorite ? PRIMARY_PURPLE : BLACK}
-        strokeWidth="2.5"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  </button>
-);
